fix conda-env tests
