+++ conda/common/yaml.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
+
 from logging import getLogger
 
+from .compat import PY2, odict
 from .._vendor.auxlib.decorators import memoize
 
 log = getLogger(__name__)
@@ -21,6 +23,31 @@ def get_yaml():
     return yaml
 
 
+yaml = get_yaml()
+
+
+def represent_ordereddict(dumper, data):
+    value = []
+
+    for item_key, item_value in data.items():
+        node_key = dumper.represent_data(item_key)
+        node_value = dumper.represent_data(item_value)
+
+        value.append((node_key, node_value))
+
+    return yaml.nodes.MappingNode(u'tag:yaml.org,2002:map', value)
+
+
+yaml.representer.RoundTripRepresenter.add_representer(odict, represent_ordereddict)
+
+if PY2:
+    def represent_unicode(self, data):
+        return self.represent_str(data.encode('utf-8'))
+
+
+    yaml.representer.RoundTripRepresenter.add_representer(unicode, represent_unicode)  # NOQA
+
+
 def yaml_load(string):
     yaml = get_yaml()
     return yaml.load(string, Loader=yaml.RoundTripLoader, version="1.2")
@@ -28,7 +55,6 @@ def yaml_load(string):
 
 def yaml_dump(object):
     """dump object to string"""
-    yaml = get_yaml()
     return yaml.dump(object, Dumper=yaml.RoundTripDumper,
                      block_seq_indent=2, default_flow_style=False,
                      indent=2)
