+++ conda_env/env.py
@@ -11,6 +11,7 @@ from conda.core.linked_data import linked
 from copy import copy
 from itertools import chain
 
+from conda_env.yaml import dump
 from . import compat, exceptions, yaml
 from .pip_util import add_pip_installed
 
@@ -163,7 +164,7 @@ class Environment(object):
 
     def to_yaml(self, stream=None):
         d = self.to_dict()
-        out = compat.u(yaml.dump(d, default_flow_style=False))
+        out = compat.u(dump(d))
         if stream is None:
             return out
         stream.write(compat.b(out, encoding="utf-8"))
